pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hdlproject"
version = "2.0.0"
description = "A modular system for managing Vivado FPGA projects"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "InquirerPy>=0.3.4",
    "prompt-toolkit>=3.0.0",
    "psutil>=5.9.0",
    "colorama>=0.4.6", 
    "rich>=13.0.0",
    "GitPython>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "hdldepends>=0.1.3",
]

[project.scripts]
hdlproject = "hdlproject.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["hdlproject*"]

[tool.setuptools.package-data]
hdlproject = [
    "tcl/*.tcl",
]